requirements.txt
PySide6==6.9.3
Pillow==10.4.0